% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblas_get_num_threads.r
\name{openblas_get_num_threads}
\alias{openblas_get_num_threads}
\title{openblas_get_num_threads}
\usage{
openblas_get_num_threads()
}
\value{
The number of threads (an integer).
}
\description{
Get the number of threads OpenBLAS is using.
}
