% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctl.r
\name{openblas_set_num_threads}
\alias{openblas_set_num_threads}
\title{openblas_set_num_threads}
\usage{
openblas_set_num_threads(num_threads)
}
\arguments{
\item{num_threads}{An integer number of threads to use.}
}
\value{
Irrelevant.
}
\description{
Set the number of threads for OpenBLAS to use.
}
\details{
Operates via side-effects.
}

