% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblasctl-package.r
\docType{package}
\name{openblasctl-package}
\alias{openblasctl-package}
\title{openblasctl}
\description{
A micro-package for controlling the number of threads used
by OpenBLAS from R.  If you're somehow incapable of using
the OMP_NUM_THREADS environment variable, then you can
use this.
}
\author{
Drew Schmidt
}
\keyword{package}

