% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblas_set_num_threads.r
\name{openblas_set_num_threads}
\alias{openblas_set_num_threads}
\title{openblas_set_num_threads}
\usage{
openblas_set_num_threads(num_threads)
}
\arguments{
\item{num_threads}{An integer number of threads to use.}
}
\value{
An invisible \code{NULL}.
}
\description{
Set the number of threads for OpenBLAS to use. Operates via side-effects.
}
\details{
OpenBLAS stores the number of threads to use as a global variable. It will
read the value of the environment variable \code{OMP_NUM_THREADS} (if one is
provided) when the calling process (in this case R) starts, and thereafter
ignore it. This means you can't simply call R's \code{Sys.setenv()} to
modify \code{OMP_NUM_THREADS}. Nor can you call the C OpenMP function
\code{omp_set_num_threads()} (at least at the time of writing) to update this
global value. So for example, the CRAN package OpenMPController will not work
with OpenBLAS.
}
