% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openblasctl-package.r
\docType{package}
\name{openblasctl-package}
\alias{openblasctl-package}
\title{openblasctl}
\description{
A micro-package for controlling the number of threads used by OpenBLAS from
R. This can be useful if you did not set the `OMP_NUM_THREADS` environment
variable before starting your R session. Obviously this requires that you are
using the OpenBLAS library with R.
}
\author{
Drew Schmidt
}
\keyword{package}
